requirements.txt
PyYAML~=6.0.2
Flask~=3.1.0
mysql-connector-python~=9.2.0